refactor to make code review easier
